Rename and fix docs
