+++ src/ios/addToHeaderSearchPaths.js
@@ -1,4 +1,4 @@
-const mapHeaderSearchPaths = require('./headerSearchPathIter');
+const mapHeaderSearchPaths = require('./mapHeaderSearchPaths');
 
 module.exports = function addToHeaderSearchPaths(project, path) {
   mapHeaderSearchPaths(project, searchPaths => searchPaths.concat(path));
