+++ src/ios/mapHeaderSearchPaths.js
@@ -1,6 +1,6 @@
 /**
  * Given Xcode project and path, iterate over all build configurations
- * and append new path to HEADER_SEARCH_PATHS section
+ * and execute func with HEADER_SEARCH_PATHS from current section
  *
  * We cannot use builtin addToHeaderSearchPaths method since react-native init does not
  * use $(TARGET_NAME) for PRODUCT_NAME, but sets it manually so that method will skip
@@ -8,7 +8,10 @@
  *
  * To workaround that issue and make it more bullet-proof for different names,
  * we iterate over all configurations and look if React is already there. If it is,
- * we assume user will need newly added dependency headers in it as well.
+ * we assume we want to modify that section either
+ *
+ * Important: That function mutates `buildSettings` and it's not pure thus you should
+ * not rely on it's return value
  */
 module.exports = function headerSearchPathIter(project, func) {
   const config = project.pbxXCBuildConfigurationSection();
