+++ src/ios/removeFromHeaderSearchPaths.js
@@ -1,4 +1,4 @@
-const mapHeaderSearchPaths = require('./headerSearchPathIter');
+const mapHeaderSearchPaths = require('./mapHeaderSearchPaths');
 
 /**
  * Given Xcode project and absolute path, it makes sure there are no headers referring to it
